Fix up open region setting html files with the new http server.

+++ Aurora/Framework/ClientInterfaces/RegionInfoForEstateMenuArgs.cs
@@ -39,7 +39,7 @@ namespace Aurora.Framework
         public int pricePerMeter;
         public int redirectGridX;
         public int redirectGridY;
-        public uint regionFlags;
+        public ulong regionFlags;
         public string regionType;
         public byte simAccess;
         public string simName;
